[Fix] [GGFE-102] fix infinite recall & kakao link
