+++ hooks/useAxiosResponse.ts
@@ -1,17 +1,30 @@
 import { AxiosError, AxiosResponse } from 'axios';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { instance } from 'utils/axios';
 import Cookies from 'js-cookie';
+import { loginState } from 'utils/recoil/login';
+import { useSetRecoilState } from 'recoil';
 
 export default function useAxiosResponse() {
+  const setLogin = useSetRecoilState(loginState);
+  const token = Cookies.get('refresh_token');
+  const [isRecalling, setIsRecalling] = useState(false);
+
   const errorResponseHandler = async (error: AxiosError) => {
-    if (error.response && error.response.status === 401) {
+    if (error.response && error.response.status === 401 && !isRecalling) {
+      setIsRecalling(true);
       const refreshToken = Cookies.get('refresh_token');
-      const res = await instance.post(
-        `/pingpong/users/accesstoken?refreshToken=${refreshToken}`
-      );
-      localStorage.setItem('42gg-token', res.data.accessToken);
-      return instance.request(error.config);
+      try {
+        const res = await instance.post(
+          `/pingpong/users/accesstoken?refreshToken=${refreshToken}`
+        );
+        localStorage.setItem('42gg-token', res.data.accessToken);
+        setIsRecalling(false);
+        return instance.request(error.config);
+      } catch (error) {
+        setIsRecalling(false);
+        return Promise.reject(error);
+      }
     } else {
       return Promise.reject(error);
     }
@@ -21,6 +34,12 @@ export default function useAxiosResponse() {
     (response: AxiosResponse) => response,
     errorResponseHandler
   );
+
+  useEffect(() => {
+    if (!token) {
+      setLogin(false);
+    }
+  }, [token]);
 
   useEffect(() => {
     return () => {
